AoC FAQ & Bug Hunter Role Fix
Included a link to the AoC 202 FAQ reddit post on request. This is in the "faq" cog alongside `!vpn` and can be used by `!aocfaq`.

The issue with the Bug Hunter role being removed is that the rolesync cog is looking for the API to respond with the user's level being "998" where the actual response is "999".